chore(mv): Moving tests to their own directory
    Made utils scaffold for tests
