+++ gdcdictionary/__init__.py
@@ -1,1 +1,3 @@
 from python import GDCDictionary, gdcdictionary
+import os
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
